Created seperate state for each dialog and removed box-shadow of the dialog
