+++ src/components/TeamMembers/TeamMemberStyle.js
@@ -99,4 +99,5 @@ const styles =makeStyles(theme => ({
         }
     })
 );
+
 export default styles;
